feat: Precompile localized pages
This PR sets up `generateStaticParams` to tell Next.js which locales are
going to be requested. This allows static pages to be compiled ahead of
time, which will drastically improve load times for visitors of our
site.

To do this, I had to move a few `"use client"`s around, but there are a
couple pages that manage their own state directly instead of doing it in
a component. These pages can't be statically generated until that's
changed.