feat: Convert (most) selects to shadcn/ui
This converts the...interesting...select designs to shadcn/ui. I'm not
entirely sure why each page seemed to have its own custom way of doing
select menus, but it's fixed now in any case.

The filter dropdown on the resources page still needs to be updated, but
other than that, dropdown-type inputs are all good.