feat: Add hooks to get single events/resources by ID
These hooks use the cache of the functions to get full event/resource
lists, so they should be pretty performant.

The `useResource` function will be particularly useful to be able to
show a resource modal for a resource that may not be part of the
already-fetched results, such as if the user was redirected to the
resources page from the homepage.