feat(resources): Use `useResource` in resource modal
When pagination is implemented, this will allow the modal to display
resources that have not yet been paginated. As mentioned in earlier
commits, if the resource happens to already have been fetched (e.g. if
the user opened the modal on a resource they browsed to manually), then
the existing cache will be used to avoid unnecessary network traffic.